requirements.txt
Flask>=2.3.3
Flask-CORS>=4.0.0
Flask-SQLAlchemy>=3.0.5
requests>=2.31.0
google-generativeai>=0.3.2
transformers>=4.35.2
sentence-transformers>=2.2.2
numpy>=1.26.0
pandas>=2.1.0
gitpython>=3.1.40
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.6
faiss-cpu>=1.7.4
PyPDF2>=3.0.1
markdown>=3.5.1
beautifulsoup4>=4.12.2
tiktoken>=0.5.1
openai>=1.3.7
huggingface-hub>=0.19.4
python-multipart>=0.0.6
Pillow>=10.1.0
